src/utils/db_connection.py
import os
from sqlalchemy import create_engine
from dotenv import load_dotenv

load_dotenv()

class DBManager:
    def __init__(self):
        self.user = os.getenv("DB_USER")
        self.password = os.getenv("DB_PASS")
        self.host = os.getenv("DB_HOST")
        self.port = os.getenv("DB_PORT")
        self.dbname = os.getenv("DB_NAME")

    def get_engine(self):
        
        url = f"postgresql://{self.user}:{self.password}@{self.host}:{self.port}/{self.dbname}"
        return create_engine(url)