requirements.txt
kafka-python
psycopg2-binary
sqlalchemy
pandas
streamlit
python-dotenv
plotly
dbt-postgres
fastf1